Fix type mismatch warning.
Reviewed-on: http://gerrit.aldebaran.lan:8080/4133
Tested-by: llec <llec@aldebaran-robotics.com>

+++ libqimessaging/tests/test_event_connect.cpp
@@ -32,7 +32,7 @@ void testDelete(bool afirst, bool disconnectFirst)
   unsigned int linkId = a.connect(fireId, &b, onFireId);
   a.connect(fireId, &b, onFireId2);
   std::vector<qi::MetaEvent::Subscriber> subs = a.metaObject().event(fireId)->subscribers();
-  EXPECT_EQ(2, subs.size());
+  EXPECT_EQ(static_cast<unsigned int>(2), subs.size());
   // Subs ordering is unspecified
   EXPECT_EQ(subs[0].method + subs[1].method, onFireId + onFireId2);
   a.emitEvent("fire", 12);
